+++ src/apis/auth/index.ts
@@ -1,13 +1,13 @@
-import Axios from "@/apis/@core/instance";
-import AuthAxios from "@/apis/@core/authInstance";
+import Axios from '@/apis/@core/instance';
+import AuthAxios from '@/apis/@core/authInstance';
 import {
   ISocialLoginResponse,
   LoginDataType,
   SocialLoginTokenType,
-} from "@/types/auth";
+} from '@/types/auth';
 
 export async function login({ email, password }: LoginDataType) {
-  const response = await Axios.post("/api/auth/login", {
+  const response = await Axios.post('/api/auth/login', {
     email,
     password,
   });
@@ -16,13 +16,13 @@ export async function login({ email, password }: LoginDataType) {
 }
 
 export async function refresh() {
-  const response = await AuthAxios.post("/api/auth/refresh");
+  const response = await AuthAxios.post('/api/auth/refresh');
 
   return response.data;
 }
 
 export async function sendVerifyEmail(email: string) {
-  const response = await Axios.post("/api/auth/send-verification-code", {
+  const response = await Axios.post('/api/auth/send-verification-code', {
     email,
   });
 
@@ -36,7 +36,7 @@ export async function sendVerifyCode({
   email: string;
   code: string;
 }) {
-  const response = await Axios.post("/api/auth/verify-email", {
+  const response = await Axios.post('/api/auth/verify-email', {
     email,
     code,
   });
@@ -48,7 +48,7 @@ export async function sendSocialLoginToken({
   code,
   provider,
 }: SocialLoginTokenType) {
-  const response = await Axios.post<ISocialLoginResponse>("/api/auth/oauth2", {
+  const response = await Axios.post<ISocialLoginResponse>('/api/auth/oauth2', {
     code,
     provider,
   });
@@ -62,7 +62,7 @@ export async function resetPassword({
   emailAuthToken: string;
   newPassword: string;
 }) {
-  const response = await AuthAxios.post("/api/auth/reset-password", {
+  const response = await AuthAxios.post('/api/auth/reset-password', {
     emailAuthToken,
     newPassword,
   });
