+++ src/pages/auth/login/index.tsx
@@ -1,9 +1,11 @@
-import LoginForm from "@/components/auth/Form/LoginForm";
+import LoginForm from '@/components/auth/Form/LoginForm';
+import SocialLoginForm from '@/components/auth/Form/SocialLoginForm';
 
 const LoginPage = () => {
   return (
     <main>
       <LoginForm />
+      <SocialLoginForm />
     </main>
   );
 };
